requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0  # for JWT handling
passlib[bcrypt]==1.7.4           # for password hashing
python-multipart==0.0.6          # for form data parsing
redis==5.0.1
email-validator==2.1.0.post1     # for email validation
PyJWT==2.8.0                     # Add this for JWT
bcrypt==4.0.1